+++ src/tests/CountryList.test.js
@@ -1,14 +1,14 @@
 import { act } from 'react';
 import ReactDOM from 'react-dom/client';
-import Listings from './Listings';
+import CountryList from '../screens/CountryList';
 
 globalThis.IS_REACT_ACT_ENVIRONMENT = true;
 
-it('renders Listings component', async () => {
+it('renders CountryList component', async () => {
   const container = document.createElement('div');
   document.body.appendChild(container);
 
   await act(async () => {
-    ReactDOM.createRoot(container).render(<Listings />);
+    ReactDOM.createRoot(container).render(<CountryList />);
   });
 });
