+++ src/tests/Listings.test.js
@@ -1,14 +1,14 @@
 import { act } from 'react';
 import ReactDOM from 'react-dom/client';
-import CountryList from './CountryList';
+import Listings from '../screens/Listings';
 
 globalThis.IS_REACT_ACT_ENVIRONMENT = true;
 
-it('renders CountryList component', async () => {
+it('renders Listings component', async () => {
   const container = document.createElement('div');
   document.body.appendChild(container);
 
   await act(async () => {
-    ReactDOM.createRoot(container).render(<CountryList />);
+    ReactDOM.createRoot(container).render(<Listings />);
   });
 });
