+++ src/tests/ListingsBody.test.js
@@ -1,14 +1,14 @@
 import { act } from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import ListingsBody from '../screens/ListingsBody';
 
 globalThis.IS_REACT_ACT_ENVIRONMENT = true;
 
-it('renders App component', async () => {
+it('renders ListingsBody component', async () => {
   const container = document.createElement('div');
   document.body.appendChild(container);
 
   await act(async () => {
-    ReactDOM.createRoot(container).render(<App />);
+    ReactDOM.createRoot(container).render(<ListingsBody />);
   });
 });
