+++ src/tests/ListingsHeader.test.js
@@ -1,14 +1,14 @@
 import { act } from 'react';
 import ReactDOM from 'react-dom/client';
-import ListingsBody from './ListingsBody';
+import ListingsHeader from '../screens/ListingsHeader';
 
 globalThis.IS_REACT_ACT_ENVIRONMENT = true;
 
-it('renders ListingsBody component', async () => {
+it('renders ListingsHeader component', async () => {
   const container = document.createElement('div');
   document.body.appendChild(container);
 
   await act(async () => {
-    ReactDOM.createRoot(container).render(<ListingsBody />);
+    ReactDOM.createRoot(container).render(<ListingsHeader />);
   });
 });
