+++ src/tests/Splash.test.js
@@ -1,14 +1,14 @@
 import { act } from 'react';
 import ReactDOM from 'react-dom/client';
-import ListingsHeader from './ListingsHeader';
+import Splash from '../screens/Splash';
 
 globalThis.IS_REACT_ACT_ENVIRONMENT = true;
 
-it('renders ListingsHeader component', async () => {
+it('renders Splash component', async () => {
   const container = document.createElement('div');
   document.body.appendChild(container);
 
   await act(async () => {
-    ReactDOM.createRoot(container).render(<ListingsHeader />);
+    ReactDOM.createRoot(container).render(<Splash />);
   });
 });
